waveview: Add a WaveView class to abstract wave view
A wave view is part of sample to be shown in a view of certain
width and height. The content of a view is determined by
the samples and the starting point of the view in sample coordinate.
The result of get_view is ready to print to command line.